requirements.txt
pygame==2.0.1
djitellopy2==2.3
opencv-python==4.5.1.48